+++ app/src/main/java/com/example/basecomposemvvm/utils/FormatCurrency.kt
@@ -3,8 +3,7 @@ package com.example.basecomposemvvm.utils
 import java.text.NumberFormat
 import java.util.Locale
 
-fun formatCurrency(amount: Long): String {
+fun formatCurrency(amount: Double): String {
     val formatter = NumberFormat.getCurrencyInstance(Locale("vi", "VN"))
-
     return formatter.format(amount)
 }
